# Improve collision cache heap allocation