--- a/dependencies_unstable.repos
+++ b/dependencies_unstable.repos
@@ -1,0 +1,16 @@
+- git:
+    local-name: ros_industrial_cmake_boilerplate
+    uri: https://github.com/ros-industrial/ros_industrial_cmake_boilerplate.git
+    version: master
+- git:
+    local-name: tesseract
+    uri: https://github.com/ros-industrial-consortium/tesseract.git
+    version: master
+- git:
+    local-name: opw_kinematics
+    uri: https://github.com/Jmeyer1292/opw_kinematics.git
+    version: master
+- git:
+    local-name: ifopt
+    uri: https://github.com/ethz-adrl/ifopt.git
+    version: master